feat : Base62 기본 문자열 인코딩
* Base62EncodingStrategy, Base62EncodingStrategyTest
1 ~ 2147483646 까지 인코딩(Integer 한계)

* ShortenedUrl
id 값 정수 범위 Integer로 변경

+++ src/main/java/com/prgrms/shortener/domain/EncodingStrategy.java
@@ -2,5 +2,5 @@ package com.prgrms.shortener.domain;
 
 public interface EncodingStrategy {
 
-  String encode(Long id);
+  String encode(Integer id);
 }
