+++ src/main/java/com/prgrms/shortener/domain/JpaShortenedUrlRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface JpaShortenedUrlRepository extends JpaRepository<ShortenedUrl, Long>, ShortenedUrlRepository {
+public interface JpaShortenedUrlRepository extends JpaRepository<ShortenedUrl, Integer>, ShortenedUrlRepository {
 
 }
