+++ src/main/java/com/prgrms/shortener/domain/ShortenedUrl.java
@@ -18,13 +18,13 @@ public class ShortenedUrl {
   @Id
   @Column
   @GeneratedValue(strategy = GenerationType.SEQUENCE)
-  private Long id;
+  private Integer id;
   @Column(unique = true)
   private String originalUrl;
   @Column(length = 7, unique = true)
   private String shortenedKey;
 
-  public ShortenedUrl(Long id) {
+  public ShortenedUrl(Integer id) {
     this.id = id;
   }
 
@@ -32,12 +32,9 @@ public class ShortenedUrl {
     this.originalUrl = url;
   }
 
-  // Controller에서 출력 데이터를 확실하게 정한다.
-  // 서버의 domain을 알고 있는 영역은 presentation layer
   public void assignKey(String key) {
     checkArgument(key.length() == 7, "key는 7개의 문자로 구성되어야 합니다.");
     this.shortenedKey = key;
   }
 
-  // 알고리즘을 enum 형태로 보관환 뒤에 entity에서 호출할 수 있도록 할까?
 }
